Update the Strobe pin usage to be inline with APB3 spec
For cases when the Requester does not support the Strobe pins, but the
Completer does - all bits of the Strobe signal should be tied to the
same value as Write.  In this case, where we don't know in advance
what the write width is, tie all bits to Write.

Signed-off-by: Edward Connell <Edward.Connell@microchip.com>
(cherry picked from commit dbccc6f8f3e586b55e575fb47e5bcfd90ebdd9bb)